Создает вспомогательный элемент City
